Changed the port

+++ src/test/java/tz/go/moh/him/thscp/mediator/ffars/muse/MediatorMainTest.java
@@ -24,7 +24,7 @@ public class MediatorMainTest {
         MediatorConfig mediatorConfig = (MediatorConfig) loadConfigMethod.invoke(null, "src/test/resources/mediator.properties");
 
         Assert.assertEquals("localhost", mediatorConfig.getServerHost());
-        Assert.assertEquals(new Integer(3021), mediatorConfig.getServerPort());
+        Assert.assertEquals(new Integer(3022), mediatorConfig.getServerPort());
         Assert.assertEquals(new Integer(60000), mediatorConfig.getRootTimeout());
         Assert.assertTrue(mediatorConfig.getHeartsbeatEnabled());
     }
